Modified the project structure.
- Created `core.py`, `__main__.py` and `__init__.py`.
- Removed `pixel_puzzle.py`, `python.ico` and `key.npz`.
- Modified `README.md` and `requirements.txt`.
- Added `pyproject.toml` and `setup.py`.
- Added `python-publish.yml`.
