+++ pyproject.toml
@@ -1,0 +1,36 @@
+[build-system]
+requires = ["setuptools>=61", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "PixelPuzzle"
+description = "Encode/decode images using Base64 or shuffle/recover the pixels of images."
+requires-python = ">=3.10"
+license = { text = "GNU GPLv3" }
+authors = [{ name = "ZhanZiyuan", email = "ziyuanzhan@mail.nankai.edu.cn" }]
+maintainers = [{ name = "ZhanZiyuan", email = "ziyuanzhan@mail.nankai.edu.cn" }]
+classifiers = [
+    "Environment :: Console",
+    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
+    "Operating System :: Microsoft :: Windows",
+    "Operating System :: POSIX :: Linux",
+    "Programming Language :: Python :: 3",
+    "Topic :: Multimedia :: Graphics",
+    "Topic :: Security :: Cryptography",
+]
+dependencies = ["numpy>=2.2.4", "pillow>=11.1.0"]
+dynamic = ["version", "readme"]
+
+[project.urls]
+"Homepage" = "https://github.com/ZhanZiyuan/PixelPuzzle"
+"Bug Tracker" = "https://github.com/ZhanZiyuan/PixelPuzzle/issues"
+
+[project.scripts]
+pixelpuzzle = "pixelpuzzle.__main__:main"
+
+[tool.setuptools]
+packages = ["pixelpuzzle"]
+
+[tool.setuptools.dynamic]
+version = { attr = "pixelpuzzle.__version__" }
+readme = { file = ["README.md"], content-type = "text/markdown" }
