+++ requirements.txt
@@ -1,2 +1,2 @@
-numpy==2.2.4
-pillow==11.1.0
+numpy>=2.2.4
+pillow>=11.1.0
